+++ Sources/VCore/Extensions/UIKit/UI View/UIView+RoundCorners.swift
@@ -37,8 +37,8 @@ extension UIView {
     view.roundCorners(.layerAllCorners, by: 10)
 
     NSLayoutConstraint.activate([
-        view.constraintWidth(to: nil, constant: 100),
-        view.constraintHeight(to: nil, constant: 100)
+        view.widthAnchor.constraint(equalToConstant: 100),
+        view.heightAnchor.constraint(equalToConstant: 100)
     ])
 
     return view