+++ Sources/VCore/Views/UIKit/Inner Shadow UI View/InnerShadowUIView.swift
@@ -90,8 +90,8 @@ open class InnerShadowUIView: UIView {
     view.backgroundColor = UIColor.systemBlue.withAlphaComponent(0.67)
 
     NSLayoutConstraint.activate([
-        view.constraintWidth(to: nil, constant: 100),
-        view.constraintHeight(to: nil, constant: 100)
+        view.widthAnchor.constraint(equalToConstant: 100),
+        view.heightAnchor.constraint(equalToConstant: 100)
     ])
 
     return view